Show buses when bus stop is clicked
TODO: need to add `estimate` endpoint data to sidebar

There is no easy way to connect the bus IDs (VehicleNo's) to the buses in the `estimates` API endpoint. So I think we should keep the sidebar as just a list of estimates for a particular stop, and use the buses on the map to report and view capacity.